--- a/base/dsfs/structure.go
+++ b/base/dsfs/structure.go
@@ -8,6 +8,15 @@ import (
 	"github.com/qri-io/qfs/cafs"
 )
 
+// BaseTabularSchema is the base schema for tabular data
+var BaseTabularSchema = map[string]interface{}{
+	"type": "array",
+	"items": map[string]interface{}{
+		"type":  "array",
+		"items": []interface{}{},
+	},
+}
+
 // loadStructure assumes path is valid
 func loadStructure(ctx context.Context, store cafs.Filestore, path string) (st *dataset.Structure, err error) {
 	data, err := fileBytes(store.Get(ctx, path))