--- a/cmd/testdata/movies/tf_modified.star
+++ b/cmd/testdata/movies/tf_modified.star
@@ -1,0 +1,4 @@
+# transform that sets a simple body
+def transform(ds, ctx):
+  result = [1,2,3]
+  ds.set_body(result)