If a body is small enough, inline it before calling friendly. This is because friendly uses deepdiff, which is not necessarily fast enough for large bodies.

Read transform scripts to see if they changed between commits, letting us have a commit where the transform changes but body stays the same.

Drop derived values from structure before making the friendly message. This is currently less than ideal, since component needs some work to operate on maps.

Diff the head and body separately to get much better estimates for how much of a body changes.

Remove orphaned test private key in the old dataset tests, move it to testPeers.