@dustmop 2020-03-26
Fixes https://github.com/qri-io/qri/issues/1217 and https://github.com/qri-io/qri/issues/1215.