@dustmop 2020-03-27
Should also mention, this is the sequel to https://github.com/qri-io/qri/pull/1031.