@dustmop 2020-03-31
Performance numbers for this feature. Note that the implementation cuts off body diffing at 20M, so we should expect larger save operations to not be affected.

I used the `.circleci/user_benchmarks.sh` script for this. Timing of `save` action, numbers are for "user","system","total", 4 runs each:

Before (master):

```
Small (10,000):
0.477 0.116 0.578
0.495 0.114 0.782
0.498 0.110 0.594
0.489 0.112 0.580

Medium (50,000):
2.167 0.339 1.952
2.135 0.331 2.010
2.117 0.315 1.886
2.107 0.319 1.911

Large (100,000):
5.050 0.736 4.320
4.240 0.595 3.606
4.290 0.608 3.643
4.294 0.603 3.638
```

After (friendly-commit):

```
Small (10,000):
0.689 0.138 0.796
0.684 0.131 0.783
0.686 0.139 0.799
0.676 0.134 0.791

Medium (50,000):
2.129 0.331 1.918
2.119 0.322 1.892
2.171 0.344 1.950
2.141 0.332 1.920

Large (100,000):
4.584 0.665 3.949
4.489 0.678 3.851
4.380 0.640 3.788
4.290 0.628 3.666
```

I also tried increasing the limit to 200M to see how it affects saving, but only ran 1 time not 4:

```
Small (10,000):
0.682 0.143 0.806

Medium (50,000):
3.254 0.447 3.066

Large (100,000):
6.594 0.956 6.115
```

We can see that including body diffing increasing the time to save by about 50%. I would argue that this is acceptable for small bodies (below 20M) as done in the implementation, but probably too slow after that.