@dustmop 2020-03-31
Something I forgot to mention, that I added as a TODO and am also copying here. Limiting the body size for diffs causes an undesirable side-effect: if a user has a dataset body larger than this size, then any `save` operation will create a commit, even if no changes have been made, since our change detection will see BodyTooBig and assume something changed. Some solutions to this:
* above 20M start using a less-accurate non-structured checksum instead of deepdiff
* calculate a structured checksum and compare that to the last commit's checksum
* immitate "redo" (https://apenwarr.ca/log/20181113), store file attributes, see if any change
* at a large enough size, like 4G, just assume the file is always changed. don't be expensive
We should make this algorithm agree with how `status` works. See issue: https://github.com/qri-io/qri/issues/1150