@b5 2020-03-31
hrm... I didn't catch the proviso for files over `BigFile`. sounds like a job for a rolling hash function.